Fixed autolayout margings (orientation / devices).
AutoLayoutMarginsViewController is now compatible with any screen size in any orientation. All the squares have a 1:1 constrains and have equal widths as the centre square. The centre square's width is fixed having its constraint connected to an IBOutlet. The constant value is then updated in `viewDidLayoutSubviews` when the size of the container view has changed.