Updates fetch-dependencies.sh to validate-environment.sh
As discussed with @orj, build user usually doesn't have rights to brew install. This requires the buildbox to have swiftlint installed for now. Hopefully we'll find a better solution in the future to manage build tools dependencies without including the binary in the repo.

Btw swiftlint isn't just a binary, there seem to be a framework to move as well :(